requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
Click==7.0
colorama==0.4.1
drive-cli==2.1.0
et_xmlfile==2.0.0
google==3.0.0
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==2.8
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pick==0.6.4
prettytable==0.7.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyDrive==1.3.1
pyfiglet==0.8.post1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.21.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.6
tzdata==2024.2
uritemplate==3.0.1
urllib3==1.24.3
windows-curses==2.4.0
